Rename storage repository to client
