+++ src/domain/instance/usecases/ListInstancesUseCase.ts
@@ -3,7 +3,7 @@ import { UseCase } from "../../common/entities/UseCase";
 import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { Instance, InstanceData } from "../entities/Instance";
 
 export interface ListInstancesUseCaseProps {
