+++ src/domain/metadata/usecases/SetResponsiblesUseCase.ts
@@ -7,7 +7,7 @@ import { Instance } from "../../instance/entities/Instance";
 import { ReceivedPullRequestNotification } from "../../notifications/entities/PullRequestNotification";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { MetadataResponsible } from "../entities/MetadataResponsible";
 
 export class SetResponsiblesUseCase implements UseCase {
