+++ src/domain/modules/usecases/DeleteModuleUseCase.ts
@@ -5,7 +5,7 @@ import { Instance } from "../../instance/entities/Instance";
 import { BasePackage, Package } from "../../packages/entities/Package";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 
 export class DeleteModuleUseCase implements UseCase {
     constructor(private repositoryFactory: RepositoryFactory, private localInstance: Instance) {}
