+++ src/domain/modules/usecases/GetModuleUseCase.ts
@@ -3,7 +3,7 @@ import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Instance } from "../../instance/entities/Instance";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { MetadataModule } from "../entities/MetadataModule";
 import { BaseModule, Module } from "../entities/Module";
