+++ src/domain/modules/usecases/ListModulesUseCase.ts
@@ -5,7 +5,7 @@ import { Instance } from "../../instance/entities/Instance";
 import { InstanceRepositoryConstructor } from "../../instance/repositories/InstanceRepository";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { MetadataModule } from "../entities/MetadataModule";
 import { BaseModule, Module } from "../entities/Module";
