+++ src/domain/modules/usecases/SaveModuleUseCase.ts
@@ -6,7 +6,7 @@ import { Instance } from "../../instance/entities/Instance";
 import { InstanceRepositoryConstructor } from "../../instance/repositories/InstanceRepository";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { Module } from "../entities/Module";
 
 export class SaveModuleUseCase implements UseCase {
