+++ src/domain/notifications/usecases/UpdatePullRequestStatusUseCase.ts
@@ -8,7 +8,7 @@ import { InstanceRepositoryConstructor } from "../../instance/repositories/Insta
 import { MetadataResponsible } from "../../metadata/entities/MetadataResponsible";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import {
     PullRequestStatus,
     ReceivedPullRequestNotification,
