+++ src/domain/package-import/usecases/ListImportedPackagesUseCase.ts
@@ -4,7 +4,7 @@ import { RepositoryFactory } from "../../common/factories/RepositoryFactory";
 import { Instance } from "../../instance/entities/Instance";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { ImportedPackageData } from "../entities/ImportedPackage";
 
 type ListImportedPackageError = "UNEXPECTED_ERROR";
