+++ src/domain/packages/usecases/DiffPackageUseCase.ts
@@ -5,7 +5,7 @@ import { Instance } from "../../instance/entities/Instance";
 import { MetadataPackage } from "../../metadata/entities/MetadataEntities";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { getMetadataPackageDiff, MetadataPackageDiff } from "../entities/MetadataPackageDiff";
 import { Store } from "../../stores/entities/Store";
 import { GitHubRepositoryConstructor } from "../repositories/GitHubRepository";
