+++ src/domain/packages/usecases/PublishStorePackageUseCase.ts
@@ -7,7 +7,7 @@ import { Instance } from "../../instance/entities/Instance";
 import { BaseModule } from "../../modules/entities/Module";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { GitHubError } from "../entities/Errors";
 import { BasePackage } from "../entities/Package";
 import { Store } from "../../stores/entities/Store";
