+++ src/domain/stores/usecases/GetStoreUseCase.ts
@@ -1,10 +1,10 @@
 import { UseCase } from "../../common/entities/UseCase";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepository } from "../../storage/repositories/StorageRepository";
+import { StorageClient } from "../../storage/repositories/StorageClient";
 import { Store } from "../entities/Store";
 
 export class GetStoreUseCase implements UseCase {
-    constructor(private storageRepository: StorageRepository) {}
+    constructor(private storageRepository: StorageClient) {}
 
     public async execute(id: string): Promise<Store | undefined> {
         const store = this.storageRepository.getObjectInCollection<Store>(Namespace.STORES, id);
