+++ src/domain/synchronization/usecases/CreatePullRequestUseCase.ts
@@ -14,7 +14,7 @@ import {
 } from "../../notifications/entities/PullRequestNotification";
 import { Repositories } from "../../Repositories";
 import { Namespace } from "../../storage/Namespaces";
-import { StorageRepositoryConstructor } from "../../storage/repositories/StorageRepository";
+import { StorageRepositoryConstructor } from "../../storage/repositories/StorageClient";
 import { SynchronizationType } from "../entities/SynchronizationType";
 
 interface CreatePullRequestParams {
